stateless: Refactor pNext, flag and enum checking utils
Move most of the low level checking methods into a separate class.
These checks are used with in the pNext checks and in many cases
require knowledge about which extensions are enabled in the VkInstance,
VkPhysicalDevice or VkDevice. The new separate object can be set
up to use the correct DeviceExtensions object and then the
extension checks within the rest of the code become much simpler.