layers: Simplify instance vs device extension handling
DeviceExtensions has always inherited from InstanceExtensions,
so having separate device_extension and instance_extension
members in ValidationObject was redundant. There's now a single
extensions member and there are 3 use cases:

1) In vvl::dispatch::Instance and ValidationObjects representing
a VkInstance, extensions represents what was in VkInstanceCreateInfo
2) In vvl::dispatch::Device and ValidationObjects representing
a VkDevice, extensions represents what was in VkDeviceCreateInfo
for device extensions, as well as what is in the VkInstance
3) In StatelessValidation objects representing VkInstances,
the physical_device_extensions map represents the supported
extensions for all physical devices.