syncval: Split Device and Instance ValidationObjects
This is the start of splitting class ValidationObject into
separate Device and Instance classes. Syncval has nothing happening
at Instance level so it gets to go first. Once all the leaf
classes are split, ValidationStateTracker and ValidationObject
can be split. This seems backwards but allows it to happen
incrementally rather than in 1 giant PR