build: Remove usage of deprecated std::aligned_storage/union
These are deprecated in c++23.

Note: in the one instance where the default Align value was used
for std::aligned_storage<Len, Alignment> (i.e.
std::aligned_storage<Len>), this patch now adopts
alignof(std::max_align_t). According to [1] that should be
a behavior change, but in practice it doesn't seem to be when
using clang nor gcc. In any case, any change in the alignment
would be an increase, which is fine.

[1] https://en.cppreference.com/w/cpp/types/aligned_storage
[2] https://godbolt.org/z/K5xGcxhjh