layers: Remove Device arg from various submit-time callbacks
Many of the callbacks don't use it. Those that do should just
capture the device in their lambda. Eventually the way these
callbacks get run will change and so that they'll be run by
the state tracker on behalf of each validation object so
the device arg will be wrong.