layers: Readd removed state creation abstractions
The whole point of these abstractions is to enable
subclasses to provide customized creation to customized
subclasses of the state objects themselves, so it would be
bad design to remove this ability. For example, this is not
only used by layer subclasses such as the best practices
layers, but also by the Vulkan SC validation layers to
ammend SC-specific data to the state objects through
subclassing.

Having overrideable factory methods for the state
objects is paramount to enable extensibility.