tests: Add helper for sync2 barriers
The helpers allow to skip VkDependencyInfo part of initialization
when barrier command specifies a single barrier, which is a common
case in the tests.

For now, there are no helpers for sync1 barriers. They can't be as nice
as sync2 helpers, because you also need two additional parameter for
the stage masks, so the signature will need 4 parameters in total and
it's not obvious if this brings benefits comparing to using sync1
api directly. We can reevaluate this in the next test improvement
iteration.