layers: Implement strict rules for DeviceWait with maintenance1
When swapchain maintenance1 is enabled, QueueWaitIdle/DeviceWaitIdle
cannot retire present semaphores and presentation fence should be
used.

For pre-maintenance1 there is no other options in the context of
halting all device activity, so QueueWaitIdle and DeviceWaitIdle is the
only option to rely on and validation does not complain about this.