layers: Add top/bottom  to the list of graphics meta stages
So far it worked correctly even without this change because generated
code lists top/bottom meta stages for graphics pipe and
syncAllCommandStagesByQueueFlags() returns them. But that probably will
change in the future and generated code will report only atomic stages.

Also top/bottom are deprecated so their usage will be minimized in the
internal details and only added explicitly when it's necessary.