layers: Remove layout map compatibility check
This code existed from the beginning and after investigation the
conclusion this check is not needed. The compatibility check and the
assert is not triggered by the tests too (neither was reported in the
issues from what I remember).

It checked for aspect mask but aspect mask is constant for the same
image object and checking image object pointer does not help either -
the image object we compare either the same or diffent in case it's
aliased images, but in the latter case we _do_ want to update state
using aliased image layout state.

Another question is maybe we need to check image id here, but
compatibility check did not do that either, so we can't regress in
this regard.