layers: Rename many image layout names
The first part of changes is that I started moving types in
image_layout_map.h to represent just regular maps rather
than additional abstractions over maps (still in progress).
We have two image layout map types:
* CommandBufferImageLayoutMap - tracks layouts of image subresources
  during command buffer recording, and in addition to the current layout
  also tracks the first layout used by the command buffer
* ImageLayoutMap - directly stores layouts of image subresources

The second part is related to Registry naming. It was used to
represent per-image layout information, but this name is better
suited to represent global storage, that collects information
for all images (so that's the registry). So Registry to name
collection of layouts for all images and LayoutMap is the layouts of
subresources of specific image.