layers: Make ImageLayoutMap a regular map
The lock functionality is moved to image. This allows to use
BothRangeMap type directly as layout map. Also not all layout maps
need a lock (only for global layout tracking), so without lock inside
the map the code is cleaner for such use cases.

In case of aliased images the lock is shared in the same way as layout
map is shared - shared_ptr. In the previous implementation the lock
existed in each image, and in case of aliasing some locks were unused,
so there was a chance to use a wrong lock.