layers: Rename initial layout to first layout
This is to avoid confusion, as we already use the term "initial layout"
to refer both to image's initial layout
(VkImageCreateInfo::initialLayout) and also initialLayout specified in
various render pass structures.

The first layout is the first detected layout of the image subresource
in the command buffer.

The new term has a matching "first access" syncval naming, which means
that we track the first memory access in the command buffer to use
later for submit-time or CmdExecuteCommands validation.