bp: Remove meta stage checks
There are scenarios when ALL_COMMANDS is the only option to use
(for example to sync some operations that do not happen on
pipeline stages).

Also it's possible to provide exampes when doing fine-grained
synchronization results in several barriers that are slower than a
single barrier that uses a meta stage.