gpuav: Fix wrong pipeline handle in error message
The pipeline handle reported by instrumentation error messages can
be incorrect due to caching of instrumented pipelines,
if the same shader is used in multiple pipelines.
Not an issue, due to a recent change the last bound pipeline is stored
in the error logging lambda and added to the LogObjectList in case of
an error.
Note: Maybe in the GPL case, all pipeline handles used to compose the
final pipeline should be logged?