sync: Introduce ApplyMarkupFunctor
This action can create new ranges according to infill_update_range
logic but otherwise does not change access state. By running this action
we get a guarantee that references to access states won't be invalidated
if we run another action over the same range. This creates an
opportunity to cache such references locally together with pending
barrier information and then apply pending barriers without storing
this information inside access state.

The goal is to implement this method of pending state management
on top of this change. This change by itself is rather a refactor to
check if ApplyMarkupFunctor works as expected.