layers: Remove VerifyImageLayoutRange
In all other places except VerifyImageLayoutRange we use
LayoutUseCheckAndMessage and ForEachMatchingLayoutMapRange directly
to implement context-specific validation.

VerifyImageLayoutRange tried to provide generic logic that can be
reused in two places, but it turns we still need a way to customize
logic depending on context. This PRs removes VerifyImageLayoutRange and
uses LayoutUseCheckAndMessage and ForEachMatchingLayoutMapRange directly
the same way as we do in other places. In a way this unifies the
implementation.

Rename Verify->Validate.
Rename ValidateImageLayoutSubresource -> ValidateSubresourceImageLayout.
Remove not need hit_error flag in video tests.