sync: Use boolean instead of infill_state pointer
Previously infill_state pointer if not null was always pointing to
default ResourceAccessState object (except one case which is handled
here). This change uses a boolean to indicate if infill is needed and
if yes, then default object is used. This is mostly to simplify further
work so I can be sure only default object is used for infill operation.