scripts: Cleanup how update_deps gets binaries
When slang was added as a dependency, the logic needed to select the
correct binary artifact to download was hardcoded in update_deps. Now,
the url is parameterized on the name of the release, the operating
system, and the architecture being used.

Doing so required refactoring update_deps to include the new fields in
the repo objects:
 - release: The name of the release to download. This is added as an
   exclusive option group with commit, so repos can be either source git
   repos or binary artifact repos.
 - url_format_map: Map of names that are used to turn python's
   platform.system() and `arch`` command line argument into strings that
   correspond to the correct binary artifact for the current system.
 - build_architectures: If present, will allow this dependency to be
   ignored on architectures that aren't supported.

Removed the --pointer_size command line argument as that is redundant
with the --arch command line argument.