sync: Clone range_map to become AccessMap
This will evolve separately based on syncval needs and will become an
AccessMap type.

The generic containers/range_map.h is used in various places but for
syncval it is also a core data structure and it's becoming increasingly
important to change it (easily) similar to other syncval code.

There are ongoing changes that are not stl-compatible and apply more
restrictions on object creation/copy/assignment to better manage
large number of syncval's access states. In general the idea is that
syncval's version will be significanlty different (potentially will
use different underlying storage and interface) to reduce allocations
and optimize accesses. Doing this with generic implementation is tricky
but also the generic implementation works pretty good as a generic
container for other usages.

One more reason why I really want this separately is because it's main
syncval container and you have to read/debug it regularly, having a
syncval specific implementation will help to simplify implementation
by removing some layers needed by generic implementation.