settings: Enable DebugOutput by default on Windows
DebugOutput is very useful on Windows. It works for the apps that do not
create console. For the apps that go with console sometimes (it depends)
it is still more convenient to work with Debug Output because that's
what Visual Studio, DebugView and other gui tools use as debug output.